extract shared YAML renderer to yaml_tree.py
Dream-origin: plan
Dream-bounces: 0
