+++ .uncoded/stubs/src/uncoded/namespace_map.pyi
@@ -1,8 +1,8 @@
 # src/uncoded/namespace_map.py
 
 from pathlib import Path
-import yaml
 from uncoded.extract import ModuleInfo
+from uncoded.yaml_tree import render_yaml_tree
 
 HEADER = ...
 
@@ -11,7 +11,3 @@ def build_map(modules: list[ModuleInfo]) -> dict:
 
 def render_map(namespace: dict) -> str:
     ...
-
-class _CleanDumper(yaml.SafeDumper):
-    def increase_indent(self, flow, indentless):
-        ...
